+++ R/add_institution.R
@@ -1,7 +1,9 @@
 
 
 #' @import data.table
+#' @importFrom checkmate qassert assert_names
 NULL
+
 
 
 #' Add a column of institution names
@@ -9,7 +11,7 @@ NULL
 #' Add a column of character values with institution names (or labels) using
 #' student ID as the join-by variable. In the MIDFIELD practice data, the
 #' labels are anonymized. Based on information in the MIDFIELD \code{term}
-#' data table.
+#' data table or equivalent.
 #'
 #' If a student is associated with more than one institution, the institution
 #' at which they completed the most terms is returned. An existing column with
@@ -30,7 +32,24 @@ NULL
 #'
 #'
 #' @examples
-#' # TBD
+#' # extract a column of IDs from student
+#' id <- toy_student[, .(mcid)]
+#'
+#'
+#' # add institutions from term
+#' DT1 <- add_institution(id, midfield_table = toy_term)
+#' head(DT1)
+#'
+#'
+#' # will overwrite institution column if present
+#' DT2 <- add_institution(DT1, midfield_table = toy_term)
+#' head(DT2)
+#'
+#'
+#'
+#'
+#'
+#'
 #'
 #'
 #' @export
@@ -39,33 +58,33 @@ NULL
 add_institution <- function(dframe,
                             midfield_table) {
 
-  # explicit arguments and NULL defaults if any
-  assert_explicit(dframe)
-  assert_explicit(midfield_table)
+  # required arguments
+  qassert(dframe, "d+")
+  qassert(midfield_table, "d+")
 
-  # check argument class
-  assert_class(dframe, "data.frame")
-  assert_class(midfield_table, "data.frame")
+  # optional arguments
+  # NA
 
-  # dframe is modified "by reference" throughout
-  setDT(dframe)
-  setDT(midfield_table)
+  # inputs modified (or not) by reference
+  dframe <- copy(as.data.table(dframe)) #  must copy
+  setDT(midfield_table) # immediately subset, so side-effect OK
 
-  # existence of required columns
-  assert_required_column(dframe, "mcid")
-  assert_required_column(midfield_table, "mcid")
-  assert_required_column(midfield_table, "institution")
-  assert_required_column(midfield_table, "term")
+  # required columns
+  assert_names(colnames(dframe),
+               must.include = c("mcid"))
+  assert_names(colnames(midfield_table),
+               must.include = c("mcid", "institution", "term"))
 
   # class of required columns
-  assert_class(dframe[, mcid], "character")
-  assert_class(midfield_table[, mcid], "character")
-  assert_class(midfield_table[, institution], "character")
-  assert_class(midfield_table[, term], "character")
+  qassert(dframe[, mcid], "s+")
+  qassert(midfield_table[, mcid], "s+")
+  qassert(midfield_table[, institution], "s+")
+  qassert(midfield_table[, term], "s+")
 
   # bind names due to NSE notes in R CMD check
   N <- NULL
 
+  # do the work
   DT <- filter_match(midfield_table,
     match_to = dframe,
     by_col = "mcid",