+++ R/add_race_sex.R
@@ -2,14 +2,16 @@
 
 #' @import data.table
 #' @importFrom wrapr stop_if_dot_args
+#' @importFrom checkmate qassert assert_names
 NULL
 
 
-#' Add two columns for student race/ethnicity and sex
+#' Add one column each for student race/ethnicity and sex
 #'
 #' Add two columns of character values of students' self-reported
 #' race/ethnicity and sex using student ID as the join-by variable.
-#' Based on information in the MIDFIELD \code{student} data table.
+#' Based on information in the MIDFIELD \code{student} data table or
+#' equivalent.
 #'
 #' Existing columns with the same names as the added columns are overwritten.
 #'
@@ -28,16 +30,21 @@ NULL
 #'
 #'
 #' @examples
-#'
-#'
 #' # Add race and sex to a data frame of graduates
 #' dframe <- toy_degree[1:5, c("mcid", "cip6")]
-#' add_race_sex(dframe = dframe, midfield_table = toy_student)
+#' add_race_sex(dframe, midfield_table = toy_student)
 #'
 #'
 #' # Add race and sex to a data frame from the term table
 #' dframe <- toy_term[21:26, c("mcid", "institution", "level")]
-#' add_race_sex(dframe = dframe, midfield_table = toy_student)
+#' add_race_sex(dframe, midfield_table = toy_student)
+#'
+#'
+#' # If present, existing race and sex columns are overwritten
+#' # Using dframe from above,
+#' DT1 <- add_race_sex(dframe, midfield_table = toy_student)
+#' DT2 <- add_race_sex(DT1, midfield_table = toy_student)
+#' all.equal(DT1, DT2)
 #'
 #'
 #' @export
@@ -46,43 +53,39 @@ NULL
 add_race_sex <- function(dframe,
                          midfield_table) {
 
-  # explicit arguments, NULL defaults if any
-  assert_explicit(dframe)
-  assert_explicit(midfield_table)
+  # required arguments
+  qassert(dframe, "d+")
+  qassert(midfield_table, "d+")
 
-  # check argument class
-  assert_class(dframe, "data.frame")
-  assert_class(midfield_table, "data.frame")
+  # optional arguments
+  # NA
 
-  # The dframe argument is modified "by reference." Thus changing its value
-  # inside the function immediately changes its value in the calling frame
-  # --- a data.table feature designed for fast data manipulation,
-  # especially for data that occupies a lot of memory.
+  # inputs modified (or not) by reference
   setDT(dframe)
-  setDT(midfield_table)
+  setDT(midfield_table) # immediately subset, so side-effect OK
 
-  # existence of required columns
-  assert_required_column(dframe, "mcid")
-  assert_required_column(midfield_table, "mcid")
-  assert_required_column(midfield_table, "race")
-  assert_required_column(midfield_table, "sex")
+  # required columns
+  assert_names(colnames(dframe),
+               must.include = c("mcid"))
+  assert_names(colnames(midfield_table),
+               must.include = c("mcid", "race", "sex"))
 
   # class of required columns
-  assert_class(dframe[, mcid], "character")
-  assert_class(midfield_table[, mcid], "character")
-  assert_class(midfield_table[, race], "character")
-  assert_class(midfield_table[, sex], "character")
+  qassert(dframe[, mcid], "s+")
+  qassert(midfield_table[, mcid], "s+")
+  qassert(midfield_table[, race], "s+")
+  qassert(midfield_table[, sex], "s+")
 
+
+  # bind names due to NSE notes in R CMD check
+  # NA
+
+  # do the work
   # preserve column order except columns that match new columns
   names_dframe <- colnames(dframe)
   added_cols <- c("race", "sex")
   key_names <- names_dframe[!names_dframe %chin% added_cols]
   dframe <- dframe[, key_names, with = FALSE]
-
-  # work on the midfield_table data
-  # cols_we_want <- c("mcid", added_cols)
-  # rows_we_want <- midfield_table$mcid %chin% dframe$mcid
-  # DT <- midfield_table[rows_we_want, cols_we_want, with = FALSE]
 
   DT <- filter_match(
     dframe = midfield_table,
@@ -90,7 +93,6 @@ add_race_sex <- function(dframe,
     by_col = "mcid",
     select = c("mcid", added_cols)
   )
-
 
   # left-outer join, keep all rows of dframe
   dframe <- merge(dframe, DT, by = "mcid", all.x = TRUE)