+++ inst/tinytest/test_add_race_sex.R
@@ -1,8 +1,11 @@
 test_add_race_sex <- function() {
 
+    library("tinytest")
+    library("checkmate")
+    using("checkmate")
+
     # function arguments
-    # add_race_sex <- function(dframe,
-    #                          midfield_table)
+    # add_race_sex <- function(dframe, midfield_table)
 
 
     library(data.table)
@@ -32,18 +35,36 @@ test_add_race_sex <- function() {
                  add_race_sex(dframe = toy_degree[1:10, .(mcid, cip6)],
                               midfield_table = toy_student))
 
+    # overwrites existing race and sex columns
+    expect_equal(DT,
+                 add_race_sex(dframe = DT, midfield_table = toy_student))
 
+    # joins to itself (equivalent due to key)
+    expect_equivalent(DT, add_race_sex(DT, midfield_table = DT))
 
+    # midfield_table argument must student or equivalent
+    expect_error(add_race_sex(DT, midfield_table = toy_degree))
 
+    # ID column required in both data frames
+    missing_column <- copy(DT)[, mcid := NULL]
+    expect_error(add_race_sex(DT, missing_column))
+    expect_error(add_race_sex(missing_column, DT))
 
+    # race and sex columns required in midfield_table argument
+    missing_column <- copy(DT)[, race := NULL]
+    expect_error(add_race_sex(DT, missing_column))
+    missing_column <- copy(DT)[, sex := NULL]
+    expect_error(add_race_sex(DT, missing_column))
+
+    # arguments must be data frames
+    expect_error(add_race_sex(NULL, midfield_table = toy_student))
+    expect_error(add_race_sex(DT$mcid, midfield_table = toy_student))
+    expect_error(add_race_sex(DT, midfield_table = NULL))
+    expect_error(add_race_sex(DT, midfield_table = toy_student$race))
 
     invisible(NULL)
 }
 
 test_add_race_sex()
-
-
-
-